// packages/apsara/components/table/Table.types.tsx
type LiteralUnion<T extends U, U = string> = T | (U & Record<never, never>);

export type RaypointFilterOption = LiteralUnion<
  string & keyof typeof RaypointFilterFns
>;
export type RaypointFilterFn<TData extends Record<string, any> = {}> =
  | FilterFn<TData>
  | RaypointFilterOption;

export type RaypointAggregationOption = string &
  keyof typeof RaypointAggregationFns;
export type RaypointAggregationFn<TData extends Record<string, any> = {}> =
  | AggregationFn<TData>
  | RaypointAggregationOption;

export type RaypointColumnDef<TData extends Record<string, any> = {}> = Omit<
  ColumnDef<TData, unknown>,
  | "accessorKey"
  | "aggregatedCell"
  | "aggregationFn"
  | "cell"
  | "columns"
  | "filterFn"
  | "footer"
  | "header"
  | "id"
  | "sortingFn"
> & {
  id?: LiteralUnion<string & keyof TData>;
  header: string;
  footer?: string;
  accessorKey?: DeepKeys<TData>;
  accessorFn?: (originalRow: TData) => any;
  aggregationFn?:
    | RaypointAggregationFn<TData>
    | Array<RaypointAggregationFn<TData>>;
  filterFn?: RaypointFilterFn<TData>;
  filterVariant?:
    | "checkbox"
    | "date"
    | "date-range"
    | "multi-select"
    | "range"
    | "select"
    | "text";
  columnDefType?: "data" | "display" | "group";
  columns?: RaypointColumnDef<TData>[];
};